Some place holders before the cleanup
