+++ src/main/drivers/serial_usb_vcp.c
@@ -32,7 +32,7 @@
 #include "usbd_cdc_vcp.h"
 #include "usb_io.h"
 #elif defined(STM32H7A3xx)
-#include "vcp_hal/stm32h7a3/usbd_cdc_interface.h"
+#include "vcp_hal/stm32h7a3/usbd_cdc_if.h"
 #elif defined(STM32F7) || defined(STM32H7)
 #include "vcp_hal/usbd_cdc_interface.h"
 #include "drivers/usb_io.h"
