+++ src/main/drivers/serial_usb_vcp_stm32h7a3.c
@@ -40,13 +40,13 @@
 #include "usbd_cdc.h"
 
 #ifdef STM32H7A3xx
-#include "vcp_hal/stm32h7a3/usbd_cdc_interface.h"
+#include "vcp_hal/stm32h7a3/usbd_cdc_if.h"
 #include "vcp_hal/stm32h7a3/usbd_desc.h"
 #else
 #error "Unknown MCU"
 #endif
 
-USBD_HandleTypeDef hUsbDeviceHS;
+extern USBD_HandleTypeDef hUsbDeviceHS;
 
 #ifdef DEBUG
 static inline void Error_Handler(void)
@@ -59,35 +59,7 @@ static inline void Error_Handler(void)
 #define Error_Handler()
 #endif
 
-void MX_USB_DEVICE_Init(void)
-{
-  /* USER CODE BEGIN USB_DEVICE_Init_PreTreatment */
-
-  /* USER CODE END USB_DEVICE_Init_PreTreatment */
-
-  /* Init Device Library, add supported class and start the library. */
-  if (USBD_Init(&hUsbDeviceHS, &HS_Desc, DEVICE_HS) != USBD_OK)
-  {
-    Error_Handler();
-  }
-  if (USBD_RegisterClass(&hUsbDeviceHS, &USBD_CDC) != USBD_OK)
-  {
-    Error_Handler();
-  }
-  if (USBD_CDC_RegisterInterface(&hUsbDeviceHS, &USBD_Interface_fops_HS) != USBD_OK)
-  {
-    Error_Handler();
-  }
-  if (USBD_Start(&hUsbDeviceHS) != USBD_OK)
-  {
-    Error_Handler();
-  }
-
-  /* USER CODE BEGIN USB_DEVICE_Init_PostTreatment */
-  HAL_PWREx_EnableUSBVoltageDetector();
-
-  /* USER CODE END USB_DEVICE_Init_PostTreatment */
-}
+void MX_USB_DEVICE_Init(void);
 
 void usbVcpInitHardware(void)
  {
@@ -122,6 +94,45 @@ void usbVcpInitHardware(void)
     // END OF MX_USB_Device_Init
     //delay(100); // Cold boot failures observed without this, even when USB cable is not connected
     */
+}
+
+// interface code
+uint32_t CDC_Receive_BytesAvailable(void) {
+    // TODO: implement / remove
+    return 0;
+}
+
+uint32_t CDC_Receive_DATA(uint8_t* recvBuf, uint32_t len)
+{
+    UNUSED(recvBuf);
+    UNUSED(len);
+    return 0;
+}
+
+uint32_t CDC_Send_DATA(const uint8_t *ptrBuffer, uint32_t sendLength) {
+    UNUSED(ptrBuffer);
+    UNUSED(sendLength);
+    // TODO: check
+    //return CDC_Transmit_HS(&hUsbDeviceHS, ptrBuffer, sendLength);
+    return 0;
+}
+
+int usbIsConnected(void)
+{
+    // TODO: verify
+    return (hUsbDeviceHS.dev_state == USBD_STATE_CONFIGURED);
+}
+
+int usbIsConfigured(void)
+{
+    // TODO: verify
+    return (hUsbDeviceHS.dev_state == USBD_STATE_CONFIGURED);
+}
+
+
+uint32_t CDC_Send_FreeBytes(void) {
+    // TODO: implement
+    return 0;
 }
 
 #endif
